requirements.txt
pandas
numpy
scikit-learn
torch
matplotlib
yfinance
sqlalchemy
psycopg2-binary
fastapi
uvicorn
python-dotenv
pyyaml